+++ src/com/jec/commons/files/reflect/utils/DecoratorParser.ts
@@ -18,6 +18,8 @@ import {DecoratorPropertiesBuilder} from "./DecoratorPropertiesBuilder";
 import {ImportRef} from "./ImportRef";
 import {ImportRefParser} from "./ImportRefParser";
 import {DecoratorProperties} from "../DecoratorProperties";
+import {DecoratorType} from "./DecoratorType";
+import {UrlStringsEnum} from "../../../util/UrlStringsEnum";
 
 /**
  * A utility class that is used by <code>FilePreProcessor</code> objects to
@@ -39,16 +41,30 @@ export class DecoratorParser {
   //////////////////////////////////////////////////////////////////////////////
 
   /**
-   * The pattern used to detect decorators in a JavaScript file.
+   * The pattern used to detect member decorators in a JavaScript file.
    */
-  private static readonly JS_PATTERN:RegExp =
+  private static readonly MEMBER_PATTERN:RegExp =
                    /__decorate\s*\((?:[^)(]+|\((?:[^)(]+|\([^)(]*\))*\))*\);/gm;
+  
+  /**
+   * The pattern used to detect field decorators in a JavaScript file.
+   */
+  private static readonly FIELD_PATTERN:RegExp = /__param\(\d, (\w*).(\w*)/gm;
 
   /**
-   * The pattern used to detect decorators in a TypeScript file.
+   * The reference to the open parenthesis character (<code>(</code>).
    */
-  private static readonly TS_PATTERN:RegExp =
-    /^\s*@[a-zA-Z]+([A-z0-9]+)?\s*\((?:[^)(]+|\((?:[^)(]+|\([^)(]*\))*\))*\)/gm;
+  private static readonly OPEN_PARENTHESIS:string = "(";
+
+  /**
+   * The reference to the newline character (<code>\n</code>).
+   */
+  private static readonly NEW_LINE:string = "\n";
+  
+  /**
+   * The reference to the metadata parameters string (<code>__param</code>).
+   */
+  private static readonly PARAM:string = "__param";
 
   /**
    * The reference to the <code>ImportRefParser</code> instance used by this
@@ -56,10 +72,112 @@ export class DecoratorParser {
    */
   private static readonly PARSER:ImportRefParser = new ImportRefParser();
 
+  /**
+   * The reference to the <code>DecoratorPropertiesBuilder</code> instance used
+   * by this decorator parser.
+   */
+  private static readonly BUILDER:DecoratorPropertiesBuilder =
+                                               new DecoratorPropertiesBuilder();
+
+  //////////////////////////////////////////////////////////////////////////////
+  // Private methods
+  //////////////////////////////////////////////////////////////////////////////
+  
+  /**
+   * Returs the class path into the import refs for the specified TypeScript
+   * class reference.
+   * 
+   * @param {string} classRef the TypeScript class reference for which to return
+   *                          the valid class path.
+   * @param {Array<ImportRef>} refs an array that contains all files imports for
+   *                                the current file.
+   * @return {string} the class path for the specified parameters.
+   */
+  public static getClassPath(classRef:string, refs:ImportRef[]):string {
+    let len:number = refs.length;
+    let importRef:ImportRef = null;
+    let classPath:string = null;
+    while(len--){
+      importRef = refs[len];
+      if(importRef.ref === classRef) {
+        classPath = importRef.classPath;
+        break;
+      }
+    }
+    return classPath;
+  }
+
+  /**
+   * A visitor function that extracts field decorators from the specified file
+   * and adds them to the <code>result</code> decorators list.
+   * 
+   * @param {Array<DecoratorProperties>} result the list of decorators to
+   *                                            complete.
+   * @param {Array<ImportRef>} refs an array that contains all files imports for
+   *                                the specified file.
+   * @param {string} file the file to parse.
+   */
+  public static extractFields(result:Array<DecoratorProperties>,
+                                      refs:Array<ImportRef>, file:string):void {
+    let found:RegExpMatchArray = null;
+    let decoratorString:string = null;
+    let props:DecoratorProperties = null;
+    let classPath:string = null;
+    while(found = DecoratorParser.FIELD_PATTERN.exec(file)){
+      decoratorString = found[0].trim();
+      classPath = this.getClassPath(found[1], refs);
+      props = DecoratorParser.BUILDER.build(
+        found[2], decoratorString, classPath, DecoratorType.FIELD
+      );
+      result.push(props);
+    }
+  }
+  
+  /**
+   * A visitor function that extracts member decorators from the specified file
+   * and adds them to the <code>result</code> decorators list.
+   * 
+   * @param {Array<DecoratorProperties>} result the list of decorators to
+   *                                            complete.
+   * @param {Array<ImportRef>} refs an array that contains all files imports for
+   *                                the specified file.
+   * @param {string} file the file to parse.
+   */
+  public static extractMembers(result:Array<DecoratorProperties>,
+                               refs:Array<ImportRef>, file:string):void {
+    let found:RegExpMatchArray = null;
+    let decoratorString:string = null;
+    let props:DecoratorProperties = null;
+    let dotPos:number = -1;
+    let cursor:number = -1;
+    let stringBuffer:string = null;
+    let decoratorName:string = null;
+    let classPath:string = null;
+    while(found = DecoratorParser.MEMBER_PATTERN.exec(file)){
+      decoratorString = found[0].trim();
+      dotPos = decoratorString.indexOf(UrlStringsEnum.DOT);
+      stringBuffer = decoratorString.substring(dotPos + 1);
+      cursor = stringBuffer.indexOf(DecoratorParser.OPEN_PARENTHESIS);
+      decoratorName = stringBuffer.substring(0, cursor);
+      cursor = decoratorString.indexOf(DecoratorParser.PARAM);
+      if(cursor === -1 || cursor < decoratorString.indexOf(decoratorName)) {
+        cursor = decoratorString.indexOf(DecoratorParser.NEW_LINE);
+        classPath = this.getClassPath(
+          decoratorString.substring(cursor + 1, dotPos).trim(),
+          refs
+        );
+        props = DecoratorParser.BUILDER.build(
+          decoratorName, decoratorString, classPath, DecoratorType.MEMBER
+        );
+        result.push(props);
+      }
+    }
+  }
+
   //////////////////////////////////////////////////////////////////////////////
   // Public methods
   //////////////////////////////////////////////////////////////////////////////
-
+  
   /**
    * Parses the specified file and returns the list of decorators it contains.
    * 
@@ -67,18 +185,11 @@ export class DecoratorParser {
    * @return {Array<DecoratorProperties>} the list of decorators for the
    *                                      specified file.
    */
-  public static findDecorators(file:string):DecoratorProperties[] {
-    let builder:DecoratorPropertiesBuilder = new DecoratorPropertiesBuilder();
-    let result:DecoratorProperties[] = new Array<DecoratorProperties>();
-    let found:RegExpMatchArray = null;
-    let decoratorString:string = null;
-    let props:DecoratorProperties = null;
-    let imports:ImportRef[] = DecoratorParser.PARSER.getImports(file);
-    while(found = DecoratorParser.JS_PATTERN.exec(file)){
-      decoratorString = found[0].trim();
-      props = builder.build(decoratorString, file, imports);
-      result.push(props);
-    }
+  public static findDecorators(file:string):Array<DecoratorProperties> {
+    let result:Array<DecoratorProperties> = new Array<DecoratorProperties>();
+    let imports:Array<ImportRef> = DecoratorParser.PARSER.getImports(file);
+    this.extractMembers(result, imports, file);
+    this.extractFields(result, imports, file);
     return result;
   }
 }